+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `getMetadata()` on payment webhooks (#69)
 - `getPaymentSource()` on payment authorized, executed, settled and failed webhooks (#69)
 - `getSchemeId()` on refund executed and payout executed webhooks
+- Support for external account beneficiary on payout executed webhook
 - Support for metadata field on payout creation, retrieval and
   webhooks ([#68](https://github.com/TrueLayer/truelayer-php/pull/68))
 - Support for metadata field on refund creation, retrieval and